refactor: 댓글로직 다른 파일로 분리
